Bug fixes
Fixed a bug that broke controlls when alt tabing.
Fixed a bug that created infinite dash shadows if the player paused while dashing.
Fixed a bug that crashed the game if the player pressed a lot of buttons while the game was paused.
Changed the dash: now the gravity doesn't affect the player while dashing.